+++ client/client_test.go
@@ -6,7 +6,7 @@ import (
 	//"sync"
 	"strings"
 	"testing"
-	//"time"
+	"time"
 
 	ps "github.com/firepear/petrel/server"
 )
@@ -30,7 +30,7 @@ func TestIdleClient(t *testing.T) {
 	sn := "localhost:60606"
 
 	// stand up server
-	s, err := ps.New(&ps.Config{Addr: sn, Msglvl: "debug"})
+	s, err := ps.New(&ps.Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: server creation fail: %s", t.Name(), err)
 	}
@@ -48,7 +48,7 @@ func TestTimeoutClient(t *testing.T) {
 	sn := "localhost:60606"
 
 	// stand up server
-	s, err := ps.New(&ps.Config{Addr: sn, Msglvl: "debug"})
+	s, err := ps.New(&ps.Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: server creation fail: %s", t.Name(), err)
 	}
@@ -66,7 +66,7 @@ func TestClientNoProtohandler(t *testing.T) {
 	sn := "localhost:60606"
 
 	// stand up server
-	s, err := ps.New(&ps.Config{Addr: sn, Msglvl: "debug"})
+	s, err := ps.New(&ps.Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: server creation fail: %s", t.Name(), err)
 	}
@@ -80,7 +80,7 @@ func TestClientNoProtohandler(t *testing.T) {
 
 	// try to connect; we should get a 400; c should be nil
 	c, err := New(&Config{Addr: sn})
-	if !strings.Contains(fmt.Sprintf("%s", err), "400") {
+	if !strings.Contains(fmt.Sprintf("%s", err), "[400]") {
 		t.Errorf("%s: err should be 400 here", t.Name())
 	}
 	if c != nil {
@@ -93,7 +93,7 @@ func TestClientProtoMismatch(t *testing.T) {
 	sn := "localhost:60606"
 
 	// stand up server
-	s, err := ps.New(&ps.Config{Addr: sn, Msglvl: "debug"})
+	s, err := ps.New(&ps.Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: server creation fail: %s", t.Name(), err)
 	}
@@ -112,7 +112,7 @@ func TestClientProtoMismatch(t *testing.T) {
 
 	// try to connect; we should get a 497; c should be nil
 	c, err := New(&Config{Addr: sn})
-	if !strings.Contains(fmt.Sprintf("%s", err), "497") {
+	if !strings.Contains(fmt.Sprintf("%s", err), "[497]") {
 		t.Errorf("%s: err should be 497 here", t.Name())
 	}
 	if c != nil {
@@ -126,7 +126,7 @@ func TestClientProtoBadStatus(t *testing.T) {
 	sn := "localhost:60606"
 
 	// stand up server
-	s, err := ps.New(&ps.Config{Addr: sn, Msglvl: "debug"})
+	s, err := ps.New(&ps.Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: server creation fail: %s", t.Name(), err)
 	}
@@ -160,7 +160,7 @@ func TestClientProtoError(t *testing.T) {
 	sn := "localhost:60606"
 
 	// stand up server
-	s, err := ps.New(&ps.Config{Addr: sn, Msglvl: "debug"})
+	s, err := ps.New(&ps.Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: server creation fail: %s", t.Name(), err)
 	}
@@ -192,7 +192,7 @@ func TestClosedDispatch(t *testing.T) {
 	sn := "localhost:60606"
 
 	// stand up server
-	s, err := ps.New(&ps.Config{Addr: sn, Msglvl: "debug"})
+	s, err := ps.New(&ps.Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: server creation fail: %s", t.Name(), err)
 	}
@@ -214,7 +214,7 @@ func TestRequestTooLong(t *testing.T) {
 	sn := "localhost:60606"
 
 	// stand up server
-	s, err := ps.New(&ps.Config{Addr: sn, Msglvl: "debug"})
+	s, err := ps.New(&ps.Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: server creation fail: %s", t.Name(), err)
 	}
@@ -228,6 +228,34 @@ func TestRequestTooLong(t *testing.T) {
 	if err == nil {
 		t.Errorf("%s: should have errored on long req", t.Name())
 	}
+	c.Quit()
+}
+
+// create a server with a high-return-code handler and send a request
+// for it
+func TestClientHighRCHandler(t *testing.T) {
+	sn := "localhost:60606"
+
+	// stand up server
+	s, err := ps.New(&ps.Config{Addr: sn})
+	if err != nil {
+		t.Errorf("%s: server creation fail: %s", t.Name(), err)
+	}
+	_ = s.Register("appdefined", appDefinedHandler)
+	defer s.Quit()
+
+	c, err := New(&Config{Addr: sn})
+	if err != nil {
+		t.Errorf("%s: %s", t.Name(), err)
+	}
+	err = c.Dispatch("appdefined", []byte{})
+	if err != nil {
+		t.Errorf("%s: dispatch failed: %s", t.Name(), err)
+	}
+	if c.Resp.Status != 2222 {
+		t.Errorf("%s: status mismatch: %d %s", t.Name(), c.Resp.Status, c.Resp.Req)
+	}
+	time.Sleep(15 * time.Millisecond)
 	c.Quit()
 }
 
@@ -246,4 +274,10 @@ func protoGenericNotSuccess(payload []byte) (uint16, []byte, error) {
 // server/net.go)
 func protoError(payload []byte) (uint16, []byte, error) {
 	return 450, []byte{}, fmt.Errorf("synthetic error")
+}
+
+// appDefinedHandler is a handler that does nothing, but has a return
+// code in the app-defined range
+func appDefinedHandler(r []byte) (uint16, []byte, error) {
+	return 2222, r, nil
 }