+++ server/server_test.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	//	"log"
 	"sync"
-	"syscall"
 	"testing"
 	"time"
 
@@ -15,26 +14,17 @@ var sn = "localhost:60606"
 
 // start and stop an idle petrel server
 func TestServerNew(t *testing.T) {
-	s, err := New(&Config{Addr: sn, Msglvl: "debug"})
+	s, err := New(&Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: failed: %s", t.Name(), err)
 	}
 	s.Quit()
 }
 
-// start and halt (via signal) a server
-func TestServerSig(t *testing.T) {
-	s, err := New(&Config{Addr: sn, Msglvl: "debug"})
-	if err != nil {
-		t.Errorf("%s: failed: %s", t.Name(), err)
-	}
-	s.sig <- syscall.SIGINT
-}
-
 // test a few failure modes, for coverage
 func TestServerNewFails(t *testing.T) {
 	// fail to create
-	s, err := New(&Config{Addr: "localhost:22", Msglvl: "debug"})
+	s, err := New(&Config{Addr: "localhost:22"})
 	if s != nil {
 		t.Errorf("%s: should not have gotten a server, but did", t.Name())
 	}
@@ -43,8 +33,8 @@ func TestServerNewFails(t *testing.T) {
 	}
 
 	// create, but try to add a handler twice
-	s, err = New(&Config{Addr: sn, Msglvl: "debug"})
-	err = s.Register("PROTOCHECK", fakehandler)
+	s, err = New(&Config{Addr: sn})
+	err = s.Register("PROTOCHECK", fakeHandler)
 	if err == nil {
 		t.Errorf("%s: added Handler twice successfully", t.Name())
 	}
@@ -54,12 +44,12 @@ func TestServerNewFails(t *testing.T) {
 // handle a client connect/disconnect
 func TestServerClientConnect(t *testing.T) {
 	//synctest.Run(func() {
-	s, err := New(&Config{Addr: sn, Msglvl: "debug"})
+	s, err := New(&Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: server.New failed: %s", t.Name(), err)
 	}
 	// connlist should have zero items
-	i := lenSyncMap(&s.cl)
+	i := lenSyncMap(s.cl)
 	if i != 0 {
 		t.Errorf("%s: s.cl should have 0 len, has %d", t.Name(), i)
 	}
@@ -70,14 +60,14 @@ func TestServerClientConnect(t *testing.T) {
 	//synctest.Wait()
 
 	// connlist should now have one entry!
-	i = lenSyncMap(&s.cl)
+	i = lenSyncMap(s.cl)
 	if i != 1 {
 		t.Errorf("%s: s.cl should have len 1, has %d", t.Name(), i)
 	}
 	//synctest.Wait()
 
 	// wait until we're back at zero conns
-	for lenSyncMap(&s.cl) > 0 {
+	for lenSyncMap(s.cl) > 0 {
 		time.Sleep(1 * time.Millisecond)
 	}
 	s.Quit()
@@ -86,7 +76,7 @@ func TestServerClientConnect(t *testing.T) {
 
 // make sure many clients at once works properly
 func TestServerClientClobber(t *testing.T) {
-	s, err := New(&Config{Addr: sn, Msglvl: "debug"})
+	s, err := New(&Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: server.New failed: %s", t.Name(), err)
 	}
@@ -96,24 +86,24 @@ func TestServerClientClobber(t *testing.T) {
 	}
 	time.Sleep(5 * time.Millisecond)
 	// we should have at least 25 in the list
-	i := lenSyncMap(&s.cl)
+	i := lenSyncMap(s.cl)
 	if i < 25 {
 		t.Errorf("%s: s.cl should have many clients; have %d", t.Name(), i)
 	}
 	// test connlist len until client drops
-	for lenSyncMap(&s.cl) > 0 {
+	for lenSyncMap(s.cl) > 0 {
 		time.Sleep(2 * time.Millisecond)
 	}
 	s.Quit()
 }
 
-// start a server with a very low payload length limit and high msglvl
+// start a server with a very low payload length limit
 func TestServerSmallPayload(t *testing.T) {
-	s, err := New(&Config{Addr: sn, Xferlim: 15, Msglvl: "fatal"})
+	s, err := New(&Config{Addr: sn, Xferlim: 15})
 	if err != nil {
 		t.Errorf("%s: failed: %s", t.Name(), err)
 	}
-	s.Register("FAKE", fakehandler)
+	s.Register("FAKE", fakeHandler)
 	cc, err := pc.New(&pc.Config{Addr: sn})
 	if err != nil {
 		t.Errorf("%s: couldn't create client: %s", t.Name(), err)
@@ -154,7 +144,7 @@ func lenSyncMap(m *sync.Map) int {
 	return i
 }
 
-// fakehandler is a handler that does nothing
-func fakehandler(r []byte) (uint16, []byte, error) {
+// fakeHandler is a handler that does nothing
+func fakeHandler(r []byte) (uint16, []byte, error) {
 	return 0, []byte{}, fmt.Errorf("fake")
 }