+++ util.go
@@ -22,6 +22,8 @@ type Msg struct {
 	Req string
 	// Status is the numeric status indicator.
 	Code uint16
+	// Txt is free-form informational content
+	Txt string
 	// Err is the error (if any) passed upward as part of the Msg.
 	Err error
 }
@@ -29,24 +31,23 @@ type Msg struct {
 // Error implements the error interface for Msg, returning a nicely
 // (if blandly) formatted string containing all information present.
 func (m *Msg) Error() string {
-	if m.Err != nil {
-		return fmt.Sprintf("c:%s r:%d (%s) [%d %s] %s",
-			m.Cid, m.Seq, m.Req, m.Code, Stats[m.Code].Txt, m.Err)
+	if m.Code <= 1024 {
+		if m.Err != nil {
+			return fmt.Sprintf("c:%s r:%d (%s, %d, %s) %s : %s",
+				m.Cid, m.Seq, m.Req, m.Code, Stats[m.Code].Txt, m.Txt, m.Err)
+		} else {
+			return fmt.Sprintf("c:%s r:%d (%s, %d, %s) %s",
+				m.Cid, m.Seq, m.Req, m.Code, Stats[m.Code].Txt, m.Txt)
+		}
 	} else {
-		return fmt.Sprintf("c:%s r:%d (%s) [%d %s]",
-			m.Cid, m.Seq, m.Req, m.Code, Stats[m.Code].Txt)
+		if m.Err != nil {
+			return fmt.Sprintf("c:%s r:%d (%s, %d) %s : %s",
+				m.Cid, m.Seq, m.Req, m.Code, m.Txt, m.Err)
+		} else {
+			return fmt.Sprintf("c:%s r:%d (%s, %d) %s",
+				m.Cid, m.Seq, m.Req, m.Code, m.Txt)
+		}
 	}
-}
-
-// GenMsg creates messages and sends them to the Msgr channel, if they
-// match or exceed Conn.ML
-func (c *Conn) GenMsg(status uint16, req string, err error) {
-	// if this message's level is below the instance's level, don't
-	// generate the message
-	if Stats[status].Lvl < c.ML {
-		return
-	}
-	c.Msgr <- &Msg{c.Sid, c.Seq, req, status, err}
 }
 
 // GenId generates a SHA256 hash of the current Unix time, in